requirements.txt
streamlit==1.33.0
google-generativeai==0.3.2
python-dotenv==1.0.1